+++ src/pages/CreateHabit/forms/PenaltyForm.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Tooltip from './Tooltip';
+import Tooltip from '../utils/Tooltip';
 
 const PenaltyForm = ({ penalty, setPenalty }) => {
   return (
